[Feature]: fix email injection error
